GeneratorCallsFinder and CallsSearchResults classes to clarify and divide the search algorithm
