+++ discern_fwk/pattern_finders/observer_pattern_finder.py
@@ -1,6 +1,6 @@
 from typing import List
-from new_model.pattern_search_result import PatternSearchResult
-from new_model.abstract_pattern_finder import AbstractPatternFinder
+from discern_fwk.pattern_finders.result_types.pattern_search_result import PatternSearchResult
+from discern_fwk.pattern_finders.abstract_pattern_finder import AbstractPatternFinder
 
 
 class ObserverPatternFinder(AbstractPatternFinder):
