+++ discern_fwk/pattern_finders/result_types/process_result.py
