+++ discern_fwk/software_project/python_file.py
@@ -1,4 +1,4 @@
-from new_model.project_file import ProjectFile
+from discern_fwk.software_project.project_file import ProjectFile
 
 
 class PythonFile(ProjectFile):
