--- new_model/calls_search_result.py
@@ -1,7 +1,0 @@
-from new_model.process_result import ProcessResult
-
-
-class CallsSearchResult(ProcessResult):
-
-    def __init__(self, name):
-        pass
